New calibration code, first working version Scripts now match on USB product ID instead of made-up name Scripts now specify AIO_calibrate instead of AIOS_calibrate Calibration results show up in log Successful calibration is a pass, failed calibration fails the DUT

+++ Source/Spectrum.cpp
@@ -172,7 +172,7 @@ int Spectrum::getBinForHarmonic(double const fundamentalFrequency, int const har
     //int fundamentalBin = roundDoubleToInt(fundamentalFrequency / frequencyResolution);
 
     // Calc target frequency ...
-    harmonicFrequency = harmonicOrder * fundamentalFrequency;
+    harmonicFrequency = (float)(harmonicOrder * fundamentalFrequency);
     
     // Calc harmonic bin ...
     k = roundDoubleToInt(harmonicFrequency / frequencyResolution);
